@@ -14,65 +14,132 @@
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "lint:check": "eslint \"{src,apps,libs,test}/**/*.ts\"",
+    "typecheck": "tsc --noEmit",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json"
+    "test:e2e": "jest --config ./test/jest-e2e.json",
+    "test:unit": "jest --testPathPattern=src/.*\\.spec\\.ts$",
+    "test:integration": "jest --testPathPattern=test/integration/.*\\.spec\\.ts$",
+    "test:coverage:ci": "jest --coverage --ci --watchAll=false --passWithNoTests",
+    "migration:create": "typeorm migration:create",
+    "migration:generate": "typeorm migration:generate -d src/config/typeorm.config.ts",
+    "migration:run": "typeorm migration:run -d src/config/typeorm.config.ts",
+    "migration:revert": "typeorm migration:revert -d src/config/typeorm.config.ts",
+    "migration:show": "typeorm migration:show -d src/config/typeorm.config.ts",
+    "schema:drop": "typeorm schema:drop -d src/config/typeorm.config.ts",
+    "schema:sync": "typeorm schema:sync -d src/config/typeorm.config.ts",
+    "seed": "ts-node src/database/seeds/run-seeds.ts",
+    "seed:dev": "NODE_ENV=development npm run seed",
+    "seed:prod": "NODE_ENV=production npm run seed",
+    "cache:clear": "ts-node -e \"require('./src/shared/cache/cache.service').flush()\"",
+    "db:reset": "npm run schema:drop && npm run migration:run && npm run seed",
+    "docker:build": "docker build -t task-management-api .",
+    "docker:build:prod": "docker build -t task-management-api:latest .",
+    "docker:run": "docker-compose up -d",
+    "docker:dev": "docker-compose -f docker-compose.dev.yml up -d",
+    "docker:stop": "docker-compose down",
+    "docker:logs": "docker-compose logs -f",
+    "docker:push": "docker push ghcr.io/your-username/task-management-api:latest",
+    "k8s:deploy": "kubectl apply -f k8s/",
+    "k8s:delete": "kubectl delete -f k8s/",
+    "k8s:logs": "kubectl logs -f deployment/task-management-api -n task-management",
+    "k8s:port-forward": "kubectl port-forward service/task-management-api-service 3000:80 -n task-management",
+    "health:check": "curl -f http://localhost:3000/health",
+    "performance:test": "k6 run tests/performance/load-test.js",
+    "security:scan": "npm audit && snyk test",
+    "build:prod": "npm run lint:check && npm run typecheck && npm run test:coverage:ci && npm run build",
+    "deploy": "./scripts/deploy.sh",
+    "dev:setup": "./scripts/dev.sh"
   },
   "dependencies": {
-    "@adminjs/express": "^4.1.0",
-    "@adminjs/nestjs": "^4.0.1",
-    "@adminjs/typeorm": "^2.0.4",
+    "@nestjs/bull": "^10.0.1",
+    "@nestjs/cache-manager": "^2.1.1",
     "@nestjs/class-transformer": "^0.4.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/config": "^2.2.0",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/jwt": "^9.0.0",
-    "@nestjs/mapped-types": "*",
-    "@nestjs/passport": "^9.0.0",
-    "@nestjs/platform-express": "^8.0.0",
-    "@nestjs/swagger": "^6.1.2",
-    "@nestjs/typeorm": "^9.0.1",
-    "@nestjsi/class-validator": "^0.2.2",
-    "add": "^2.0.6",
-    "adminjs": "^5.9.3",
-    "bcrypt": "^5.0.1",
+    "@nestjs/class-validator": "^0.13.4",
+    "@nestjs/common": "^10.2.10",
+    "@nestjs/config": "^3.1.1",
+    "@nestjs/core": "^10.2.10",
+    "@nestjs/event-emitter": "^2.0.3",
+    "@nestjs/jwt": "^10.2.0",
+    "@nestjs/mapped-types": "^2.0.4",
+    "@nestjs/passport": "^10.0.2",
+    "@nestjs/platform-express": "^10.2.10",
+    "@nestjs/platform-socket.io": "^10.2.10",
+    "@nestjs/schedule": "^4.0.0",
+    "@nestjs/swagger": "^7.1.16",
+    "@nestjs/terminus": "^10.2.0",
+    "@nestjs/throttler": "^5.1.1",
+    "@nestjs/typeorm": "^10.0.1",
+    "@nestjs/websockets": "^10.2.10",
+    "bcrypt": "^5.1.1",
+    "bull": "^4.12.2",
+    "cache-manager": "^5.3.2",
+    "cache-manager-redis-store": "^3.0.1",
     "class-transformer": "^0.5.1",
-    "config": "^3.3.7",
-    "express": "^4.18.1",
-    "express-formidable": "^1.2.0",
-    "express-session": "^1.17.3",
-    "passport": "^0.6.0",
-    "passport-jwt": "^4.0.0",
-    "pg": "^8.8.0",
+    "class-validator": "^0.14.0",
+    "compression": "^1.7.4",
+    "config": "^3.3.9",
+    "cookie-parser": "^1.4.6",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "express-rate-limit": "^7.1.5",
+    "handlebars": "^4.7.8",
+    "helmet": "^7.1.0",
+    "ioredis": "^5.3.2",
+    "joi": "^17.11.0",
+    "mime-types": "^2.1.35",
+    "multer": "^1.4.5-lts.1",
+    "nodemailer": "^6.9.7",
+    "passport": "^0.7.0",
+    "passport-jwt": "^4.0.1",
+    "passport-local": "^1.0.0",
+    "pg": "^8.11.3",
+    "redis": "^4.6.10",
     "reflect-metadata": "^0.1.13",
-    "rimraf": "^3.0.2",
-    "rxjs": "^7.2.0",
-    "typeorm": "^0.3.7",
-    "webpack": "^5.74.0"
+    "rimraf": "^5.0.5",
+    "rxjs": "^7.8.1",
+    "sharp": "^0.33.1",
+    "socket.io": "^4.7.4",
+    "swagger-ui-express": "^5.0.0",
+    "typeorm": "^0.3.17",
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.0.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.0.0",
-    "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^4.28.2",
-    "@typescript-eslint/parser": "^4.28.2",
-    "eslint": "^7.30.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "jest": "^27.0.6",
-    "prettier": "^2.3.2",
-    "supertest": "^6.1.3",
-    "ts-jest": "^27.0.3",
-    "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "^3.10.1",
-    "typescript": "^4.3.5"
+    "@nestjs/cli": "^10.2.1",
+    "@nestjs/schematics": "^10.0.3",
+    "@nestjs/testing": "^10.2.10",
+    "@types/bcrypt": "^5.0.2",
+    "@types/bull": "^4.10.0",
+    "@types/compression": "^1.7.5",
+    "@types/cookie-parser": "^1.4.6",
+    "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.8",
+    "@types/mime-types": "^2.1.4",
+    "@types/multer": "^1.4.11",
+    "@types/node": "^20.9.0",
+    "@types/nodemailer": "^6.4.14",
+    "@types/passport-jwt": "^3.0.13",
+    "@types/passport-local": "^1.0.38",
+    "@types/pg": "^8.10.9",
+    "@types/supertest": "^2.0.16",
+    "@types/uuid": "^9.0.7",
+    "@typescript-eslint/eslint-plugin": "^6.12.0",
+    "@typescript-eslint/parser": "^6.12.0",
+    "eslint": "^8.54.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "jest": "^29.7.0",
+    "prettier": "^3.1.0",
+    "source-map-support": "^0.5.21",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.2.2"
   },
   "jest": {
     "moduleFileExtensions": [
