@@ -1,19 +1,21 @@
-require('dotenv').config();
+import {
+    DataSource
+} from 'typeorm';
+import {
+    config
+} from 'dotenv';
 
-module.exports = {
-    name: 'default',
-    type: dbConfig.type,
-    host: process.env.DB_HOST,
-    port: process.env.DB_PORT,
+config(); // Load environment variables
+
+export default new DataSource({
+    type: 'postgres',
+    host: process.env.DB_HOST || 'localhost',
+    port: parseInt(process.env.DB_PORT, 10) || 5432,
     username: process.env.DB_USERNAME,
     password: process.env.DB_PASSWORD,
     database: process.env.DB_DATABASE,
-    entities: ['dist/**/*.entity{ .ts,.js}'],
-    synchronize: true,
-    // migrations: ['dist/db/migrations/*{.ts,.js}'],
-    // migrationsTableName: 'migrations',
-    // migrationsRun: false,
-    // cli: {
-    //     migrationsDir: 'src/db/migrations',
-    // },
-};
+    entities: ['src/**/*.entity{.ts,.js}'],
+    migrations: ['src/database/migrations/*{.ts,.js}'],
+    migrationsTableName: 'migrations',
+    synchronize: false,
+});
